+++ src/App.jsx
@@ -4,18 +4,35 @@ import * as S from './App.style';
 import { GlobalStyles } from './GlobalStyles';
 import AppRoutes from './routes';
 import UserContext from './context'
-import { useDispatch } from 'react-redux';
-import { setCurrentTrack } from './store/redux/playerSlice';
+import { useDispatch, useSelector } from 'react-redux';
+import { setCurrentTrack, setCurrentTrackList } from './store/redux/playerSlice';
 import { resetAuth } from './store/redux/authSlice';
+import { selectorCategoryTrackId, selectorFavorites, selectorTrackList } from './store/selectors/selectors';
 
 function App() {
   const dispatch = useDispatch();
   const [user, setUser] = useState(JSON.parse(localStorage.getItem('user')));
   const [currentTodo, setCurrentTodo] = useState(null);
+  const favoritesTrack = useSelector(selectorFavorites);
+    const categoryTrackId = useSelector(selectorCategoryTrackId);
+    const trackList = useSelector(selectorTrackList);
 
   const handleTodoClick = (todo) => {
     setCurrentTodo(todo);
     dispatch(setCurrentTrack(todo))
+        if (location.pathname === "/") {
+      dispatch(setCurrentTrackList(trackList))
+    }
+    if (location.pathname === "/myplaylist") {
+      dispatch(setCurrentTrackList(favoritesTrack));
+    }
+    if (
+      location.pathname === `/category/1` ||
+      location.pathname === `/category/2` ||
+      location.pathname === `/category/3`
+    ) {
+      dispatch(setCurrentTrackList(categoryTrackId))
+    }
   }
 
   const navigate = useNavigate();
